refactor(images): moved subcomponent forms under SelectUpstreamImages
