+++ src/app/images/components/SelectUpstreamImages/components/SelectUpstreamImageSourcesForm/SelectedImagesTable/index.ts
