+++ src/app/images/components/SelectUpstreamImages/components/SelectUpstreamImageSourcesForm/SelectedImagesTable/useSelectedImagesTableColumns/useSelectedImagesTableColumns.tsx
