+++ src/app/images/components/SelectUpstreamImages/components/SelectUpstreamImagesForm/SelectUpstreamImagesSelect/_index.scss
